refactor: moved header components to own folder; moved header menu to own component
