+++ client/src/App.js
@@ -1,6 +1,6 @@
 /* eslint-disable react/react-in-jsx-scope */
 import './App.css';
-import Header from './components/Header';
+import Header from './components/Header/Header';
 import CreateProfilePage from './pages/CreateAccountPage';
 
 function App() {
